requirements.txt
flask==2.0.1
openai==1.3.0
python-dotenv==0.19.0
requests==2.28.1
Pillow==9.2.0
